fixed #61 Under certain circumstances the offset returned from the pairwise alignment of two DR sequences would cause an invalid index inside ReadHolder::updateStartStops().
During the alignment in Aligner::alignSlave() there was no checking
to make sure that the aligned bases were not some internal motif
that may be common to the DR group or including sequences that were
close but not really from the DR.  This seemed to be a major problem
in DR groups that were 'bad' and would get thrown out anyway.  However
I have noticed in some of the smaller datasets that many groups have
a few less spacers in them than they had before, I believe that this
is a good thing (ie those 'spacers' were not real).